+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Tail lined data with inline structs is parsed correctly.
+
 ## [0.4.2] - 2023-11-07
 
 ### Added